Fix read tap device failed when OS wake up from sleep.
Adjust the structure of the source file to the header file.
